anomaly/requirements.txt
gradio
ultralytics
requests
opencv-python
numpy
pillow